moving gray-badge to a different folder
