--- app/components/gray-badge/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./gray-badge";
